requirements-dev.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
black==21.5b0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
decorator==5.0.7
greenlet==1.1.0
html5lib==1.1
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lxml==4.6.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
parso==0.8.2
pathspec==0.8.1
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
pylama==7.7.1
pyparsing==2.4.7
pytest==6.2.4
pytest-pycodestyle==2.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
selenium==3.141.0
six==1.16.0
snowballstemmer==2.1.0
SQLAlchemy==1.4.15
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
webdriver-manager==3.4.1
webencodings==0.5.1